Added coloured end rods
By combining an end rod and a dye, you get that coloured end rod. Right clicking on a coloured end rod will set its base to that colour. Unlike the candles, this uses block states, because (1) these are whole new blocks, so thats easy to do, and (2) these need to change block textures, not just particles, so block entities would not work for it.